Inventory Manager Part 3: Incorporate Edit and Deleting of Listing
