+++ src/components/InventoryManager/index.js
@@ -1,0 +1,3 @@
+import InventoryManagerScreen from './InventoryManagerScreen';
+
+export default InventoryManagerScreen;
