+++ src/components/Marketplace/MarketplaceScreen.js
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react';
 import Airtable from 'airtable';
-import './MarketplaceScreen.css';
 import { makeStyles } from '@material-ui/core/styles';
 import FarmCard from './FarmCard';
