+++ src/components/Marketplace/index.js
@@ -1,4 +1,3 @@
-import InventoryManagerScreen from './InventoryManagerScreen';
 import MarketplaceScreen from './MarketplaceScreen';
 
-export { InventoryManagerScreen, MarketplaceScreen };
+export default MarketplaceScreen;
